remove multi_client test files into test dir to reuse code
